+++ src/components/Filters/index.tsx
@@ -1,0 +1,17 @@
+import { FiltersButton } from "./styles";
+
+export function Filters() {
+  return (
+    <ul>
+      <li>
+        <FiltersButton to={"all"}>All</FiltersButton>
+      </li>
+      <li>
+        <FiltersButton to={"active"}>Active</FiltersButton>
+      </li>
+      <li>
+        <FiltersButton to={"completed"}>Completed</FiltersButton>
+      </li>
+    </ul>
+  );
+}
